6.9

+++ 信息学奥赛一本通/1.语言与算法基础篇/2.基础算法/3.递推算法/1190.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int main(){
-	int f[71];
+	long long f[71];
 	memset(f,0,sizeof(f));
 	f[1] = 1;
 	f[2] = 2;
